# doc/api.yaml
components:
  schemas:
    UsernameFormat:
      description: |-
        The required format of a username
      type: string
      example: Sir Wilburt Coddington III
      pattern: '^[a-zA-Z0-9 ]+$'
      minLength: 3
      maxLength: 30   
...
  responses:  
    InvalidUsername:
      description: (400) Username invalid
      content:
        text/plain:
          schema:
            type: string
            example: (400) Username invalid
    Unauthorised:
      description: (401) Operation unauthorised, identifier missing or invalid
      content:
        text/plain:
          schema:
            type: string
            example: (401) Operation unauthorised, identifier missing or invalid
    NotFound:
      description: (404) Resource not found
      content:
        text/plain:
          schema:
            type: string
            example: (404) Resource not found
...
  securitySchemes:
    identifier:
      type: http
      scheme: bearer
      bearerFormat: JWT

###########################
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you!
  version: 1.0.0
paths:
...
  /users/{username}/profile/photos:
    parameters:
    - name: username
      in: path
      description: Current user's username
      required: True
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    - name: n-images
      in: header
      description: Number of images user wants to upload
      required: True
      schema:
        type: integer
        minimum: 0
        maximum: 9999
    post:
      tags: ["U2Self"]
      summary: Post one or more new photo(s)
      description: |
        Post new photo(s) on profile. Note that the photoID is automatically generated server-side, so client doesn't need to worry about it.   
      operationId: uploadPhoto
      requestBody:
        description: |
          Uploaded photo(s) alongside metadata
        content:
          multipart/form-data:
            schema:
              type: object
              properties: 
                n-images:
                  type: integer
                  minimum: 1
                  maximum: 10
                  description: |
                    Number of images user is uploading
                Description:
                  type: string
                  minLength: 0
                  maxLength: 500
                  description: |
                    Optional description accompanying image(s)
                UploadDate:
                  type: string
                  format: date-time
                  description: |
                      Date at which upload is performed. Generated client-sid.
                UploadedImagei:
                  type: string
                  format: binary
                  description: |
                    The image files will be contained here, and there will be as many of these fields as there are images. "i" specifies the index of each image and must be replaced by corresponding integer (i.e. UploadedImage1, UploadedImage2...)
      responses:
        "204":
          description: Images uploaded succesfully
        "400":
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
                example: Invalid header field values
        "401":
          $ref: "#/components/responses/Unauthorised"
      security:
        - identifier: []
...
  /users/{username}/profile/following/{user2}:
    parameters:
    - name: username
      in: path
      required: true
      description: Current user's username
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    - name: user2
      in: path
      required: true
      description: Searched user's username
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    put:
      tags: ["U2U"]
      summary: Follow given user
      description: |
        Add searched user to followed
      operationId: followUser
      responses:
        "204":
          description: User followed
        "400":
          $ref: "#/components/responses/InvalidUsername"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - identifier: []
    delete:
      tags: ["U2U"]
      summary: Unfollow given user
      description: |
        Remove searched user from follow list. If user has already been unfollowed the operation is considerd successful. If username isn't found it is specified in the response.
      operationId: unfollowUser
      responses:
        "204":
          description: User unfollowed
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/InvalidUsername"
...
  /users/{username}/blacklist/{user2}:
    parameters:
    - name: username
      in: path
      required: true
      description: Current user's username
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    - name: user2
      in: path
      required: true
      description: Target user's username
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    put:
      tags: ["U2U"]
      operationId: banUser
      summary: Ban target user
      description: |
        Add target user to blacklist. If target user has already been blacklisted, the operation is considered successful. If user doesn't exist (username not found) it is specified in response.  
      responses:
        "204":
          description: User blacklisted
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/InvalidUsername"
      security:
        - identifier: []
    delete:
      tags: ["U2U"]
      operationId: unbanUser
      summary: Unban target user
      description: |
        Remove target user from blacklist. Implements idempotency like the above PUT, or specifies upon failure to find target.
      responses:
        "204":
          description: User removed from blacklist
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/InvalidUsername"
      security:
        - identifier: []
  /photos/{uploader}/{photoID}/likes/{username}:
    parameters:
    - name: uploader
      in: path
      required: true
      description: Photo uploader's username
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    - name: photoID
      in: path
      required: true
      description: Liked photo's ID
      schema:
        type: integer
        minimum: 0
        maximum: 9999
    - name: username
      in: path
      required: true
      description: Current user's username
      schema:
        $ref: "#/components/schemas/UsernameFormat"
    put:
      tags: ["U2Photo"]
      summary: Like given photo
      description: |
        Add current user to list of people who like the given photo. Implements idempotency.
      operationId: likePhoto
      responses:
        "204":
          description: Photo liked
        "401":
          $ref: "#/components/responses/Unauthorised"
      security:
        - identifier: []
    delete:
      tags: ["U2Photo"]
      summary: Unlike a photo
      description: |
        Removes user from list of people who like the photo. Also implements idempotency.
      operationId: unlikePhoto
      responses:
        "204":
          description: Photo unliked
        "401":
          $ref: "#/components/responses/Unauthorised"
      security:
        - identifier: []
...