# service/api/opid-unban-user.go
package api

import (
	"net/http"
	"strings"
	"log"
	"fmt"
	"errors"
	"database/sql"

	"github.com/julienschmidt/httprouter"
)

// rt.router.DELETE("/users/:username/blacklist/:user2", rt.unbanUser)
func (rt *_router) unbanUser(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
	// authenticate user
	username := strings.TrimPrefix(ps.ByName("username"), "username=")
	user1Data, err := rt.db.GetUserData(username)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows){
			http.Error(w, "Provided username is invalid", http.StatusBadRequest)
			return
		}
		http.Error(w, "Something went wrong", http.StatusInternalServerError)
		log.Println("Error getting user data: ", err)
		return
	}
	err = validateToken(r, user1Data.UserID, rt.seckey)
	if err != nil {
		if strings.Contains(err.Error(), "unauthorized") || strings.Contains(err.Error(), "token signature is invalid"){
			w.WriteHeader(http.StatusUnauthorized)
			fmt.Fprint(w, "Operation unauthorised, identifier missing or invalid")
		} else {
			http.Error(w, "Something went wrong", http.StatusInternalServerError)
			log.Println("Error performing authorization check: ", err)
		}
		return
	}

	// get user2's data
	targetUser := strings.TrimPrefix(ps.ByName("user2"), "user2=")
	user2Data, err := rt.db.GetUserData(targetUser)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows){
			http.Error(w, "Searched user does not exist", http.StatusNotFound)
			return
		}
		http.Error(w, "Something went wrong", http.StatusInternalServerError)
		log.Println("Error getting user data: ", err)
		return
	}

	err = rt.db.UnbanUser(user1Data.UserID, user2Data.UserID)
	if err != nil {
		http.Error(w, "Something went wrong", http.StatusInternalServerError)
		log.Println("Error deleting blacklist pair: ", err)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}